server/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmdline==0.2.0
dart==1.0
DateTime==4.8
distlib==0.3.5
Django==4.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
environ==1.0
filelock==3.7.1
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
npm==0.1.1
optional-django==0.1.0
Pillow==9.3.0
pipenv==2022.12.19
platformdirs==2.5.2
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pygame==2.2.0
python-decouple==3.8
python-dotenv==0.21.0
python2==1.2
pytils==0.4.1
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
tk==0.1.0
toml==0.10.2
tools==0.1.9
urllib3==1.26.13
virtualenv==20.15.1
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.5.2
